EnemyStateMachine 구현

+++ Jam-up-Cave/Assets/Scripts/Player/PlayerAttack.cs
@@ -9,7 +9,7 @@ namespace Player
     {
         public float AttackCooldown { get; private set; }
         public float ReloadTime { get; private set; }
-        public float Damage { get; private set; }
+        public int Damage { get; private set; }
         
         public int TotalAmmo { get; private set; }
         public int CurrentAmmo { get; private set; }
